+++ src/dj_angles/mappers/mapper.py
@@ -4,7 +4,7 @@ from typing import Optional, Union
 
 from django.utils.module_loading import import_string
 
-from dj_angles.mappers.angles import map_call, map_model
+from dj_angles.mappers.angles import map_call, map_form, map_model
 from dj_angles.mappers.django import map_autoescape, map_block, map_css, map_extends, map_image
 from dj_angles.mappers.include import map_include
 from dj_angles.modules import is_module_available
@@ -32,6 +32,7 @@ TAG_NAME_TO_DJANGO_TEMPLATE_TAG_MAP: dict[Optional[str], Union[Callable, str]] =
     "css": map_css,
     "model": map_model,
     "call": map_call,
+    "form": map_form,
 }
 """Default mappings for tag names to Django template tags."""